+++ app/src-ui/src/App.tsx
@@ -2,6 +2,7 @@ import { HashRouter, Routes, Route, NavLink } from "react-router-dom";
 import { SettingsPage } from "./pages/SettingsPage";
 import { SyncPage } from "./pages/SyncPage";
 import { PuzzlePage } from "./pages/PuzzlePage";
+import { StatsPage } from "./pages/StatsPage";
 import "./App.css";
 
 function App() {
@@ -15,6 +16,7 @@ function App() {
               Sync
             </NavLink>
             <NavLink to="/puzzles">Puzzles</NavLink>
+            <NavLink to="/stats">Stats</NavLink>
             <NavLink to="/settings">Settings</NavLink>
           </div>
         </nav>
@@ -22,6 +24,7 @@ function App() {
           <Routes>
             <Route path="/" element={<SyncPage />} />
             <Route path="/puzzles" element={<PuzzlePage />} />
+            <Route path="/stats" element={<StatsPage />} />
             <Route path="/settings" element={<SettingsPage />} />
           </Routes>
         </main>